# public/openapi.yaml
openapi: 3.1.0
info:
  title: Mokkify Backend API
  description: >-
    Management API for Mokkify mock endpoints, response/relay templates,
    request logs, settings, and agent API keys. Every path under
    `/backend/*` requires a bearer credential (a user JWT from `POST
    /backend/auth`, or an API key from `POST /backend/api-key`) except
    `/backend/auth` itself.
  version: "2.0.0"
  license:
    name: MIT
    url: https://github.com/wavix/mokkify/blob/main/LICENSE
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /backend/template/{templateId}:
    parameters:
      - $ref: "#/components/parameters/templateId"
    get:
      tags: [template]
      operationId: getTemplate
      summary: Get a response template by id
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: "#/components/schemas/ResponseTemplate"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Invalid id or server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [template]
      operationId: updateTemplate
      summary: Update a response template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResponseTemplateWriteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  template:
                    $ref: "#/components/schemas/ResponseTemplate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [template]
      operationId: deleteTemplate
      summary: Delete a response template
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Not found or invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Either a user JWT from `POST /backend/auth` or an API key from
        `POST /backend/api-key` (format `<key_id>.<secret>`).
  parameters:
...
    templateId:
      name: templateId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
    Success:
      type: object
      properties:
        success:
          type: boolean
...
    ResponseTemplate:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: [string, "null"]
        body_parsed:
          description: JSON-parsed form of `body` when it is valid JSON.
        code:
          type: integer
        content_type:
          type: string
        headers:
          type: [object, "null"]
          additionalProperties:
            type: string
        user_id:
          type: integer
        created_at:
          type: string
          format: date-time
    ResponseTemplateWriteRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: [string, "null"]
        code:
          type: integer
          default: 200
        content_type:
          type: string
          pattern: "^[\\w.+-]+/[\\w.+-]+$"
        headers:
          type: [object, "null"]
          additionalProperties:
            type: string
...